Переработано приложение. Добавлен главный экран с выбором домашнего задания. 14 домашка перенесена в другую Activity Перенесена как есть, в дальнейшем может подправлю для красоты.
