+++ .idea/misc.xml
@@ -3,7 +3,9 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
-        <entry key="..\:/Users/Dmitry/AndroidStudioProjects/HomeWork/app/src/main/res/layout/activity_main.xml" value="0.22" />
+        <entry key="..\:/Users/Dmitry/AndroidStudioProjects/HomeWork/app/src/main/res/layout/activity_hw14.xml" value="0.17074275362318841" />
+        <entry key="..\:/Users/Dmitry/AndroidStudioProjects/HomeWork/app/src/main/res/layout/activity_main.xml" value="0.25" />
+        <entry key="..\:/Users/Dmitry/AndroidStudioProjects/HomeWork/app/src/main/res/layout/activity_main2.xml" value="0.22" />
         <entry key="..\:/Users/Dmitry/AndroidStudioProjects/HomeWork/app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.2075" />
       </map>
     </option>
