+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,6 @@
 <resources>
     <string name="app_name">HomeWork</string>
-    <string name="Main_text">Home Work</string>
+    <string name="Main_text">Home Work #14-#15</string>
+    <string name="madeBy">Made by Dmitry Drako (─‿‿─)</string>
     <string name="Hello">Hello everyone</string>
 </resources>
